a 3ds max file with the sponza scene, with materials set for mental ray rendering
renderSetup.ms selects the camera in the scene (Camera001) and positions/directs it according to data found in cameraValues.txt. Auto-set keyframe mode must be enabled. 

cameraValues.txt should be formated in the following manner:
first integer is the number of frames to be rendered. The following 3 numbers are the first frames camera position. The next 3 frames are that scene's camera target position.